feat(balancer): rename ewma to latency
