+++ kong/runloop/balancer/balancers.lua
@@ -112,7 +112,7 @@ local function create_balancer_exclusive(upstream)
       ["consistent-hashing"] = require("kong.runloop.balancer.consistent_hashing"),
       ["least-connections"] = require("kong.runloop.balancer.least_connections"),
       ["round-robin"] = require("kong.runloop.balancer.round_robin"),
-      ["ewma"] = require("kong.runloop.balancer.ewma"),
+      ["latency"] = require("kong.runloop.balancer.latency"),
     }
   end
